Moved a komponent
Moved a komponent from step one to step zero
