+++ src/components/StepOne.js
@@ -1,17 +1,13 @@
 import styled from 'styled-components';
 // Fancy nav Icons from nav aksel
-import { PersonFillIcon, CheckmarkCircleFillIcon } from '@navikt/aksel-icons';
+import { PersonFillIcon } from '@navikt/aksel-icons';
 
 // Styles thrue js
-import { TextBox, StepTitle, StepHeader, StepText, NextStepButton, Txt, InputBox, InputLabel, CheckMark } from "../components/styledComponents"
+import { TextBox, StepTitle, StepHeader, StepText, NextStepButton, Txt, InputBox, InputLabel } from "../components/styledComponents"
 
 const StepOne = ({stepOneRef, stepTwoRef, nextStepButton, activeStep, setActiveStep, id, setId,}) => {
 
-    const getPersonData = async () => {
-        const response = await fetch(`http://localhost:3000/user/${id}`)
-        const data = await response.json()
-        console.log(data)
-      }
+
 
 
     return(
@@ -20,7 +16,6 @@ const StepOne = ({stepOneRef, stepTwoRef, nextStepButton, activeStep, setActiveS
             <StepHeader>Steg 1 av 3</StepHeader>
             <StepTitle>Om meg</StepTitle>
             <PersonFillIcon className="stepIcon" title="a11y-title" color={activeStep >= 1 ? "blue" : "gray"} fontSize="1.5rem" />
-            {activeStep > 1 && <CheckMark><CheckmarkCircleFillIcon color='green' fontSize="2rem"></CheckmarkCircleFillIcon></CheckMark>}
             {activeStep === 1 && 
               <>
                 <Txt><b>Hei Starte Nybedrift</b></Txt>
@@ -30,12 +25,9 @@ const StepOne = ({stepOneRef, stepTwoRef, nextStepButton, activeStep, setActiveS
                   fra offentlige tjenester. Er svarene feil? 
                   Da er det bare å endre i svarboksene.  
                 </StepText>
-                <InputLabel>Fødselsnummer:</InputLabel>
-                <InputBox onChange={(e) => setId(e.target.value)}></InputBox>
                 {/* Moves the user to the next part */}
                 <NextStepButton onClick={(e) => {
                   nextStepButton(e, 2, stepTwoRef)
-                  getPersonData(id)
                 }}>Neste Steg</NextStepButton>
               </>
             }
